\name{variance}
\alias{variance}
\title{Calculates the variance of a sequence}
\description{
Calculates the variance of a sequence}
\usage{
variance(x)}
\arguments{
  \item{x}{the whole sequence}}
\value{a number as the variance of the sequence}
\examples{
variance(c(88,89,90,87,86))
variance(c(77,78,81,76,78,77,66,90))
variance(c(25,35,45,55,10))}

